classes refactorized, UI class started
