--- src/main/java/Difficult.java
@@ -1,3 +1,0 @@
-public enum Difficult {
-    EASY, MEDIUM , HARD;
-}
