+++ src/main/java/UI/CreationMenu.java
@@ -1,0 +1,15 @@
+package UI;
+
+public class CreationMenu {
+
+    public void mainCreationMenu(){
+        System.out.println( "1. Scape Room\n" +
+                            "2. Room\n" +
+                            "3. Enigma\n" +
+                            "4. Clue\n" +
+                            "5. Decoration item\n" +
+                            "6. Exit\n");
+    }
+
+        //ON PROGRESS
+}
