+++ src/main/java/model/Difficult.java
@@ -1,0 +1,5 @@
+package model;
+
+public enum Difficult {
+    EASY, MEDIUM , HARD;
+}
