+++ src/main/java/model/Material.java
@@ -1,3 +1,5 @@
+package model;
+
 public enum Material {
     WOOD, METAL, PLASTIC, PAPER, CRYSTAL;
 }
