+++ src/main/java/model/Theme.java
@@ -1,3 +1,4 @@
+package model;
 
 public enum Theme {
     MEDIEVAL_CASTLE, ABANDONED_SPACESHIP, HAUNTED_MANSION, JUNGLE_ADVENTURE, MURDER_MYSTERY;
